% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_params.R
\name{add_table_params}
\alias{add_table_params}
\title{Add parameter specific to the grob class.}
\usage{
add_table_params(d, params, fun, ...)
}
\arguments{
\item{d}{data.frame or matrix}

\item{params}{named list of table params}

\item{fun}{a function that returns a grob object.}
}
\description{
Add parameter specific to the grob class.
}
\author{
Yoann Pradat
}
\keyword{internal}
