% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable_utils.R
\name{gtable_rbind}
\alias{gtable_rbind}
\alias{gtable_cbind}
\title{rbind two or more gtables}
\usage{
gtable_rbind(..., size = "max", height = NULL, z = NULL)

gtable_cbind(..., size = "max", width = NULL, z = NULL)
}
\arguments{
\item{...}{gtables}

\item{size}{how should the widths be calculated?
\enumerate{
\item \code{max} maximum of all widths
\item \code{min} minimum of all widths
\item \code{first} widths/heights of first gtable
\item \code{last} widths/heights of last gtable
}}

\item{height}{padding height between grobs}

\item{z}{optional z level}

\item{width}{padding width between grobs}
}
\value{
a \code{gtable} object

a \code{gtable} object
}
\description{
rbind two or more gtables

cbind two or more gtables
}
\keyword{internal}
