% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable_utils.R
\name{cbind_2}
\alias{cbind_2}
\title{cbind two gtables}
\usage{
cbind_2(x, y, size = "max", width = NULL)
}
\arguments{
\item{x}{}

\item{size}{how should the widths be calculated?
\enumerate{
\item \code{max} maximum of all widths
\item \code{min} minimum of all widths
\item \code{first} widths/heights of first gtable
\item \code{last} widths/heights of last gtable
}}

\item{width}{padding width between grobs}
}
\description{
cbind two gtables
}
\keyword{internal}
