% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorgrobs.R
\name{circle_grob}
\alias{circle_grob}
\title{Vectorised version of circleGrob}
\usage{
circle_grob(
  r = unit(0.48, "npc"),
  fill = "black",
  col = "black",
  lty = "solid",
  lwd = 1,
  cex = 1,
  alpha = 1,
  lineend = "round",
  linejoin = "round",
  linemitre = 10,
  lex = 1,
  name = NULL,
  vp = NULL,
  x = 0.5,
  y = 0.5,
  default.units = "npc"
)
}
\value{
a \code{grob} from the grid package.
}
\description{
Vectorised version of circleGrob
}
\keyword{internal}
