% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_extra.R
\name{ttheme_awesome}
\alias{ttheme_awesome}
\title{Define theme for awesome table plot.}
\usage{
ttheme_awesome(
  base_size = 8,
  base_colour = "black",
  base_family = "",
  core_size = unit(10, "mm"),
  n_cat = 10,
  pal = "black",
  pal_breaks = NULL,
  parse = FALSE,
  padding = unit(c(0.3, 0.3), "mm"),
  ...
)
}
\arguments{
\item{base_size}{default font size}

\item{base_colour}{default font colour}

\item{base_family}{default font family}

\item{core_size}{cell size for core background grobs}

\item{n_cat}{number of size categories for core foreground grobs}

\item{pal}{color palette for core foreground grobs}

\item{pal_breaks}{breaks for color palette for core foreground grobs}

\item{parse}{logical, default behaviour for parsing text as plotmath}

\item{padding}{length-2 unit vector specifying the horizontal and vertical padding of text within each cell}

\item{...}{extra parameters added to the theme list}
}
\description{
Define theme for awesome table plot.
}
\section{Functions}{
\itemize{
\item \code{ttheme_awesome}: default theme for circle tables
}}

\author{
Yoann Pradat
}
