% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_extra.R
\name{extra_table_grob}
\alias{extra_table_grob}
\alias{ttheme_awesome}
\title{Graphical display of a textual table

Builds on top of \code{tableGrob} from gridExtra R package.}
\usage{
extra_table_grob(
  dscale,
  dcolor = NULL,
  rows = rownames(dscale),
  cols = colnames(dscale),
  rows_more = NULL,
  cols_more = NULL,
  rows_more_title = "",
  cols_more_title = "",
  theme = ttheme_awesome(),
  vp = NULL
)
}
\arguments{
\item{dscale}{a matrix}

\item{dcolor}{(optional) a matrix}

\item{rows}{(optional) a character vector}

\item{cols}{(optional) a character vector}

\item{theme}{list of theme parameters}

\item{vp}{optional viewport}
}
\value{
A gtable.
}
\description{
Create a gtable containing circle grobs representing a numeric matrix.
}
\section{Functions}{
\itemize{
\item \code{extra_table_grob}: return a grob
}}

\examples{
library(tableExtra)
}
\author{
Yoann Pradat
}
