% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtables.R
\name{gtable_legend}
\alias{gtable_legend}
\title{Build a grob containing a legend.}
\usage{
gtable_legend(
  d,
  labels,
  widths,
  heights,
  fg_fun,
  fg_params,
  bg_fun = NULL,
  bg_params = NULL,
  title_x = NULL,
  title_y = NULL,
  title_label = "Title",
  title_gp = gpar(fontsize = 10),
  labels_pad = unit(-1, "mm"),
  labels_gp = gpar(fontsize = 6),
  padding = unit(0.3, "mm"),
  name = "legend",
  vp = NULL,
  orientation = "horizontal",
  ...
)
}
\arguments{
\item{d}{data.frame or matrix}

\item{widths}{optional \code{unit.list} specifying the grob widths}

\item{heights}{optional \code{unit.list} specifying the grob heights}

\item{fg_fun}{grob-drawing function}

\item{fg_params}{named list of params passed to fg_fun}

\item{bg_fun}{grob-drawing function}

\item{bg_params}{named list of params passed to bg_fun}

\item{title_x}{\code{unit} specifying the x position of the title}

\item{title_y}{\code{unit} specifying the x position of the title}

\item{title_label}{character vector}

\item{padding}{\code{unit.list} object specifying the padding between adjacent cells.}

\item{name}{optional name of the grob}

\item{vp}{optional viewport}

\item{...}{additional parameters passed to \code{add_table_params}.}
}
\value{
A gtable.
}
\description{
Build a grob with a legend inside.
}
\author{
Yoann Pradat
}
\keyword{internal}
