% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtables.R
\name{gtable_text}
\alias{gtable_text}
\title{Build a table with text grobs.}
\usage{
gtable_text(
  d,
  widths,
  heights,
  fg_fun = text_grob,
  fg_params = list(),
  bg_fun = rect_grob,
  bg_params = list(),
  padding = unit(c(4, 4), "mm"),
  name = "table",
  vp = NULL
)
}
\arguments{
\item{d}{data.frame or matrix}

\item{heights}{optional \code{unit.list} specifying the grob heights}

\item{fg_fun}{grob-drawing function}

\item{fg_params}{named list of params passed to fg_fun}

\item{bg_fun}{grob-drawing function}

\item{bg_params}{named list of params passed to bg_fun}

\item{padding}{\code{unit.list} object specifying the padding between adjacent cells.}

\item{name}{optional name of the grob}

\item{vp}{optional viewport}

\item{width}{optional \code{unit.list} specifying the grob widths}
}
\value{
A gtable.
}
\description{
This function is a copy of the internal function \code{gtable_table} of gridExtra package.
}
\author{
Yoann Pradat
}
\keyword{internal}
