% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable_utils.R
\name{insert.unit}
\alias{insert.unit}
\title{Insert unit}
\usage{
insert.unit(x, values, after = length(x))
}
\arguments{
\item{x}{x param}

\item{values}{values param}

\item{after}{(optional) after param}
}
\description{
Helper for \link{rbind_2} and \link{cbind_2}.
}
\keyword{internal}
