% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcwag_counts-data.R
\docType{data}
\name{pcawg_counts}
\alias{pcawg_counts}
\title{Mutation counts attributed to each mutational signature in each tumour.}
\format{
A data frame with 2780 rows and 68 variables
\describe{
\item{Cancer.Types}{37 different cancer types}
\item{Sample.Names}{Unique tumour identifiers}
\item{Accuracy}{Cosine similarity between the tumour's mutational profile and the reconstructed mutational profile}
\item{SBS1}{Single-base-substitution signature 1. See \url{https://cancer.sanger.ac.uk/cosmic/signatures/SBS/SBS1.tt}}
...
}
}
\source{
Synapse collaborative compute space, \url{https://www.synapse.org/#!Synapse:syn11804065}
}
\usage{
data(pcawg_counts)
}
\description{
Data from the mutational signature analysis of Alexandrov et al. on the PCAWG data.
The data is available as supplementary data to the paper "The repertoire of mutational signatures in human cancer"
on the Synapse data repository syn11738669
}
\details{
\code{pcawg_counts} contains the mutation counts as attributed by the SigProfiler algorithm on the
Single-Base-Substitution (SBSs) mutation catalogs of 2780 WGS tumours from the PCAWG.
}
\examples{
data(pcawg_counts)
}
\references{
Alexandrov, L.B., Kim, J., Haradhvala, N.J. et al. The repertoire of mutational signatures in human cancer.
Nature 578, 94–101 (2020). \url{https://doi.org/10.1038/s41586-020-1943-3}
}
\keyword{datasets}
