% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorgrobs.R
\name{rect_grob}
\alias{rect_grob}
\title{Vectorised version of rectGrob}
\usage{
rect_grob(
  fill = "white",
  col = "black",
  lty = "solid",
  lwd = 1,
  cex = 1,
  alpha = 1,
  lineend = "round",
  linejoin = "round",
  linemitre = 10,
  lex = 1,
  name = NULL,
  vp = NULL,
  just = "centre",
  hjust = 0.5,
  vjust = 0.5,
  width = unit(1, "npc") - unit(2, "scaledpts"),
  height = unit(1, "npc") - unit(2, "scaledpts"),
  x = 0.5,
  y = 0.5,
  default.units = "npc"
)
}
\value{
a \code{grob} from the grid package.
}
\description{
Vectorised version of rectGrob
}
\keyword{internal}
