% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_params.R
\name{table_params}
\alias{table_params}
\title{Build a table of parameters.}
\usage{
table_params(d, fg_params, bg_params = NULL, rep_mode = "row")
}
\arguments{
\item{d}{data.frame or matrix}

\item{fg_params}{named list of params}

\item{bg_params}{named list of params}
}
\description{
Build a table of parameters.
}
\author{
Yoann Pradat
}
\keyword{internal}
