% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_params.R
\name{table_params_circle}
\alias{table_params_circle}
\title{Add circle parameters}
\usage{
table_params_circle(dscale, dcolor, n_cat, r_max, pal, pal_breaks)
}
\arguments{
\item{n_cat}{integer indicating the number of differently-sized circles}

\item{r_max}{unit object indicating the maximum radius}

\item{pal}{a character vector of color names}

\item{pal_breaks}{a numeric vector of break points}

\item{d}{data.frame or matrix}
}
\description{
Add circle parameters
}
\author{
Yoann Pradat
}
\keyword{internal}
