% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_params.R
\name{table_params_circle_scale}
\alias{table_params_circle_scale}
\title{Add circle scale parameters}
\usage{
table_params_circle_scale(
  d,
  scale_breaks = 10,
  r_min = unit(4, "mm"),
  r_max = unit(10, "mm"),
  d_min = NULL,
  d_max = NULL
)
}
\arguments{
\item{d}{data.frame or matrix}

\item{scale_breaks}{integer indicating the number of differently-sized circles}

\item{r_max}{unit object indicating the maximum radius}
}
\description{
Add circle scale parameters
}
\author{
Yoann Pradat
}
\keyword{internal}
