% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_params.R
\name{table_params_color}
\alias{table_params_color}
\title{Add circle color parameters}
\usage{
table_params_color(d, color_palette, color_breaks = NULL)
}
\arguments{
\item{d}{data.frame or matrix}

\item{color_palette}{a character vector of color names}

\item{color_breaks}{a numeric vector of break points}
}
\description{
Add circle color parameters
}
\author{
Yoann Pradat
}
\keyword{internal}
