% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_params.R
\name{table_params_rect}
\alias{table_params_rect}
\title{Add rect parameters}
\usage{
table_params_rect(dcolor, color_palette = NULL, color_breaks = NULL)
}
\arguments{
\item{color_palette}{a character vector of color names}

\item{color_breaks}{a numeric vector of break points}

\item{d}{data.frame or matrix}

\item{scale_breaks}{integer indicating the number of differently-sized circles}

\item{r_max}{unit object indicating the maximum radius}
}
\description{
Add rect parameters
}
\author{
Yoann Pradat
}
\keyword{internal}
