% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_params.R
\name{table_params_text}
\alias{table_params_text}
\title{Add text parameters}
\usage{
table_params_text(d)
}
\arguments{
\item{d}{data.frame or matrix}
}
\description{
Add text parameters
}
\author{
Yoann Pradat
}
\keyword{internal}
