% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableExtra.R
\docType{package}
\name{tableExtra}
\alias{tableExtra}
\title{An easy-to-use tool for drawing paper-quality tables.}
\description{
tableExtra provides a function to draw a table with grobs of varying size and colors to represent two
different types of information about multiple variables in multiple samples.
The package was originally developed to reproduce Figure 3 of Alexandrov, L.B., Kim, J., Haradhvala, N.J. et al.
The repertoire of mutational signatures in human cancer. Nature 578, 94–101 (2020).
\doi{10.1038/s41586-020-1943-3}
}
\author{
Yoann Pradat
}
