% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorgrobs.R
\name{text_grob}
\alias{text_grob}
\title{Vectorised version of textGrob}
\usage{
text_grob(
  label,
  parse = FALSE,
  col = "black",
  fontsize = 12,
  cex = 1,
  fontfamily = "",
  fontface = 1L,
  lineheight = 1.2,
  alpha = 1,
  rot = 0,
  check.overlap = FALSE,
  name = NULL,
  vp = NULL,
  just = "centre",
  hjust = 0.5,
  vjust = 0.5,
  x = 0.5,
  y = 0.5,
  default.units = "npc"
)
}
\value{
a \code{grob} from the grid package.
}
\description{
Vectorised version of textGrob
}
\keyword{internal}
