% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttheme_awesome.R
\name{ttheme_awesome}
\alias{ttheme_awesome}
\title{Define theme for awesome table plot.}
\usage{
ttheme_awesome(
  base_size = 8,
  base_colour = "black",
  base_family = "",
  core_size = unit(10, "mm"),
  scale_breaks = 10,
  scale_ratio = 0.25,
  color_palette = "black",
  color_breaks = NULL,
  rep_mode = "col",
  parse = FALSE,
  padding = unit(c(0.3, 0.3), "mm"),
  ...
)
}
\arguments{
\item{base_size}{default font size}

\item{base_colour}{default font colour}

\item{base_family}{default font family}

\item{core_size}{cell size for core background grobs}

\item{scale_breaks}{number of size categories for core foreground grobs or numeric vector of bin breaks}

\item{scale_ratio}{ratio of minimum to maximum core foreground grobs sizes}

\item{color_palette}{color palette for core foreground grobs}

\item{color_breaks}{bin breaks for color palette for core foreground grobs}

\item{rep_mode}{'col' or 'row'. Used when recycling fg_params or bg_params to make a matrix of params.}

\item{parse}{logical, default behaviour for parsing text as plotmath}

\item{padding}{length-2 unit vector specifying the horizontal and vertical padding of text within each cell}

\item{...}{extra parameters added to the theme list}
}
\description{
Define theme for awesome table plot.
}
\author{
Yoann Pradat
}
