% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_extra.R
\name{ttheme_awesome}
\alias{ttheme_awesome}
\title{Define theme for awesome table plot.}
\usage{
ttheme_awesome(
  base_size = 12,
  base_colour = "black",
  base_family = "",
  n_circle = 10,
  parse = FALSE,
  padding = unit(c(0.005, 0.005), "npc"),
  ...
)
}
\arguments{
\item{base_size}{default font size}

\item{base_colour}{default font colour}

\item{base_family}{default font family}

\item{parse}{logical, default behaviour for parsing text as plotmath}

\item{padding}{length-2 unit vector specifying the horizontal and vertical padding of text within each cell}
}
\description{
Define theme for awesome table plot.
}
\section{Functions}{
\itemize{
\item \code{ttheme_awesome}: default theme for circle tables
}}

\author{
Yoann Pradat
}
