% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_extra.R
\name{extra_table_grob}
\alias{extra_table_grob}
\alias{ttheme_awesome}
\title{Graphical display of a textual table

Builds on top of \code{tableGrob} from gridExtra R package.}
\usage{
extra_table_grob(
  d,
  rows = rownames(d),
  cols = colnames(d),
  theme = ttheme_awesome(),
  vp = NULL,
  ...
)
}
\arguments{
\item{theme}{list of theme parameters}

\item{vp}{optional viewport}

\item{...}{further arguments to control the gtable}

\item{se}{\code{SummarizedExperiment} object with rownames, colnames, rowData, colData.}
}
\value{
A gtable.
}
\description{
Create a gtable containing circle grobs representing a numeric matrix.
}
\section{Functions}{
\itemize{
\item \code{extra_table_grob}: return a grob
}}

\examples{
library(tableExtra)
}
\author{
Yoann Pradat
}
