% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_extra_grob.R
\name{table_extra_grob}
\alias{table_extra_grob}
\title{Graphical display of a table with circles of varying scales and colours.

The code is inspired by the \code{\link[gridExtra:tableGrob]{gridExtra::tableGrob()}} function.}
\usage{
table_extra_grob(
  dscale,
  dcolor = NULL,
  dscale_min = NULL,
  dscale_max = NULL,
  rows = rownames(dscale),
  cols = colnames(dscale),
  rows_more = NULL,
  cols_more = NULL,
  theme = ttheme_awesome(),
  vp = NULL
)
}
\arguments{
\item{dscale}{a matrix}

\item{dcolor}{(optional) a matrix}

\item{dscale_min}{value for setting the minimum scale size of foreground grobs. Entries in the \code{dscale} matrix
below \code{dscale_min} will have a scale of 0 (no grob).}

\item{dscale_max}{value for setting the maximum scale size of foreground grobs. Entries in the \code{dscale} matrix
above \code{dscale_max} will have a scale of 0 (no grob).}

\item{rows}{(optional) a character vector}

\item{cols}{(optional) a character vector}

\item{rows_more}{(optional) a named list of additional columns (right-part) of the plot for describing the rows. The
list names will be used as column headers.}

\item{cols_more}{(optional) a named list of additional rows (top-part) of the plot for describing the columns The
list names will be used as row headers.}

\item{theme}{list of theme parameters}

\item{vp}{optional viewport}
}
\value{
An R object of class \code{grob}
}
\description{
Create a gtable containing circle grobs representing a numeric matrix.
}
\examples{
library(tableExtra)
}
\seealso{
\code{\link[=ttheme_awesome]{ttheme_awesome()}}
}
\author{
Yoann Pradat
}
